+++ crates/ethcore/src/engines/hbbft/hbbft_engine.rs
@@ -374,9 +374,7 @@ impl IoHandler<()> for TransitionHandler {
                                     }
 
                                     let id: usize = std::process::id() as usize;
-
                                     let thread_id = std::thread::current().id();
-
                                     info!(target: "engine", "Waiting for Signaling shutdown to process ID: {id} thread: {:?}", thread_id);
 
                                     if let Some(ref weak) = *self.client.read() {
@@ -1153,7 +1151,7 @@ impl HoneyBadgerBFT {
 
     fn is_syncing(&self, client: &Arc<dyn EngineClient>) -> bool {
         match client.as_full_client() {
-            Some(full_client) => full_client.is_major_syncing(),
+            Some(full_client) => full_client.is_syncing(),
             // We only support full clients at this point.
             None => true,
         }